fix(docs): fix nested layout with page query
